goblet/requirements.txt
google-cloud-logging
google-cloud-pubsub
jsonschema
